tos-img-resize-container/requirements.txt
flask
pillow
tos